scraper/requirements.txt
requests
openai
mysql-connector-python
